requirements.txt

fastapi==0.109.2
uvicorn==0.27.1

httpx==0.26.0

pydantic==2.6.1

boto3==1.34.34

redis==5.0.1

sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.2

openai==1.12.0

google-genai>=1.0.0

confluent-kafka>=2.3.0

opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

pytest==7.4.4
pytest-asyncio==0.23.4